requirements.txt
flask
flask-cors
tensorflow-cpu
numpy
pillow
gunicorn
h5py